+++ source/Calamari.GoogleCloudAccounts/Calamari.GoogleCloudAccounts.csproj
@@ -1,0 +1,18 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <RootNamespace>Calamari.GoogleCloudAccounts</RootNamespace>
+        <TargetFrameworks>net462;netstandard2.1</TargetFrameworks>
+        <LangVersion>latest</LangVersion>
+    </PropertyGroup>
+
+    <ItemGroup Condition="'$(TargetFramework)' == 'net462'">
+        <PackageReference Include="Microsoft.Net.Http" Version="2.2.29" />
+        <Reference Include="Microsoft.CSharp" />
+    </ItemGroup>
+
+    <ItemGroup>
+        <ProjectReference Include="..\Calamari.Common\Calamari.Common.csproj" />
+    </ItemGroup>
+
+</Project>
